+++ dmr/controller.py
@@ -114,7 +114,7 @@ class Controller(Generic[_SerializerT_co], View):  # noqa: WPS214
             and validates common error responses.
         is_abstract: Whether or not this controller is abstract.
             We consider controller "abstract" when it does not have
-            exact serializer type.
+            exact serializer type or exact ``api_endpoints`` instances.
         is_async: Whether or not this controller is async.
         streaming: Does this controller work with streaming responses like SSE?
         controller_validator_cls: Runs full controller validation on definition.
@@ -181,7 +181,6 @@ class Controller(Generic[_SerializerT_co], View):  # noqa: WPS214
         if serializer is None:
             return  # this is an abstract controller
 
-        cls.is_abstract = False
         cls.serializer = serializer
         cls.settings_validator_cls(serializer=cls.serializer)()
 
@@ -193,6 +192,7 @@ class Controller(Generic[_SerializerT_co], View):  # noqa: WPS214
             )
             for canonical, meth in cls._find_existing_http_methods().items()
         }
+        cls.is_abstract = not bool(cls.api_endpoints)
         cls.is_async = cls.controller_validator_cls()(cls)
 
     @override
@@ -540,6 +540,15 @@ class Controller(Generic[_SerializerT_co], View):  # noqa: WPS214
 
     @classmethod
     def _infer_serializer(cls) -> type[_SerializerT_co] | None:
+        existing_serializer: type[_SerializerT_co] | None = getattr(
+            cls,
+            'serializer',
+            None,
+        )
+        if existing_serializer is not None:
+            # It was already set by a base class. See #873
+            return existing_serializer
+
         type_args = infer_type_args(cls, Controller)
         if not type_args:
             raise UnsolvableAnnotationsError(